requirements.txt

moviepy
gtts
pillow
ffmpeg-python
